LayoutTests/editing/deleting/delete-text-before-picture.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<div id="editor" contenteditable>Test<picture><source srcset="resources/apple.gif" media="(min-width: 600px)"/><source srcset="resources/mozilla.gif"/><img src=""></picture></div>
<script>

Markup.description('Test text deletion before <picture>.');

var picture = document.querySelector('picture');
window.getSelection().setBaseAndExtent(picture, 2, picture, 2);
Markup.dump(editor, 'Initial state');

document.execCommand('Delete');
Markup.dump(editor, 'After deletion');

</script>
</body>
</html>
