LayoutTests/editing/deleting/delete-text-before-picture-expected.txt

Test text deletion before <picture>.

Initial state:
| "Test"
| <picture>
|   <source>
|     media="(min-width: 600px)"
|     srcset="resources/apple.gif"
|   <source>
|     srcset="resources/mozilla.gif"
|   <#selection-caret>
|   <img>
|     src=""

After deletion:
| "Tes<#selection-caret>"
| <picture>
|   <source>
|     media="(min-width: 600px)"
|     srcset="resources/apple.gif"
|   <source>
|     srcset="resources/mozilla.gif"
|   <img>
|     src=""
